Fix formCreateDealSelector and add some values in dontResetSearchParams, and rename folder modules to features
