+++ getcourse/module-analytics/src/features/add-fields.js
