+++ getcourse/module-analytics/src/features/redirect-using-class.js
