+++ getcourse/module-analytics/src/features/redirect-using-code.js
