+++ getcourse/module-analytics/src/features/set-search-params.js
