+++ getcourse/module-analytics/src/features/validate-offer-settings.js
