+++ getcourse/module-analytics/src/features/values-manually.js
